xProfile: fix redirect when an admin uploads an avatar to another user's profile.
When cropping a new avatar, a redirect always took you to the logged-in user's profile.
This was fine in most situations, as a user can only update their own profile.
However, site admins are able to edit other users' profiles, so we need to redirect to the displayed user's profile.

Fixes #5580, props milesstewart88

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8322 cdf35c40-ae34-48e0-9cc9-0c9da1808c22