Bump BP version number in `package.json`.
The format for this field needs to be in the "semver" standard, which might cause issues for
future (>=3.0) or legacy (<2.0) versions of BuddyPress, if BuddyPress was ever included in any
PHP package managers. In semver, changing the first digit represents a break in compatiblilty
with previous versions. At least for now, everything remains fine. :)

Props netweb for catching this typo and for the patch, see #5160.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8391 cdf35c40-ae34-48e0-9cc9-0c9da1808c22