Core: add new `search_wildcard` parameter to `BP_User_Query`.
This new parameter controls where BuddyPress places MySQL wildcard characters around the search term (when it's set).
Prior to this change, BuddyPress added a wildcard character to both ends of the search term, but for some use cases, a more restrictive search pattern is required; you can specify "left", "right", or "both" (the default) to pick where the wildcard is added.

Fixes #5769

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8675 cdf35c40-ae34-48e0-9cc9-0c9da1808c22