Activity: fix unit tests (broken in r8697) by making sure `BP_Activity_Activity` returns integers for integer DB columns.
r8697 changed a `==` check to a stricter `===` check. When an activity item was fetched by creating a `BP_Activity_Activity` object, properties that were stored in the DB in integer columns were being returned by strings, and this caused the unit tests to fail. This change explicitly casts these properties as integers before the object is returned.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8708 cdf35c40-ae34-48e0-9cc9-0c9da1808c22