Do not add "new_member" activities if xProfile component is disabled
In Members component, bp_core_new_user_activity() automatically creates an "xProfile" activity item ("new_member" type) as soon as a user account is activated. If this happens when the xProfile component is disabled, the proper activity action callback is not available, as a result an empty string is saved in the action field of the activity table. On front-end and back-end, these mini activities are displayed without any actions.

This is a temporary fix for 2.1, we will think about something more substantial for 2.2.

See #5807

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8818 cdf35c40-ae34-48e0-9cc9-0c9da1808c22