Remove 'Edit' post link on BuddyPress pages.
Changes in bp_core_load_template() (r8932) to set the post to the current
BP directory page and #5418 to set the post type to 'page' resulted in
showing the "Edit" post link at the bottom of BuddyPress pages and in
the WP Toolbar for users with the 'edit_post' capability.

Since the "Edit" post link does not make sense on BuddyPress pages, this
commit manually removes the 'Edit' post link from rendering on BuddyPress
pages and in the WP Toolbar.

Fixes #5853.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8994 cdf35c40-ae34-48e0-9cc9-0c9da1808c22