Make sure the parent activity exists before adding a comment to it.
There can be a case when the activity displayed in the stream has been deleted by another user. Adding a comment should not be possible for this particular case as the parent activity does not exist anymore. The patch is adding a check on the {{{date_recorded}}} field of the parent activity before adding a comment. If it is not set, then false is returned and a default message is displayed to the user by {{{bp_legacy_theme_new_activity_comment()}}}. As boonebgorges, some improvements on the message can still be added.

See #5907

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9052 cdf35c40-ae34-48e0-9cc9-0c9da1808c22