During unit tests, use non-SSL mirror for Trac ticket cache when openssl is not enabled
This change is necessary to ensure that tests against open tickets are
properly skipped when running the test suite in environment where openssl is
not available (due to a change in wordpress.org infrastructure that forces all
traffic over SSL). It's primarily an issue on travis-ci's PHP 5.2 build, but
could arise in any situation where PHP was not built with openssl support.

This fallback can be removed when PHP 5.2.x support is dropped at some point in
the future.

Fixes #5910

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9053 cdf35c40-ae34-48e0-9cc9-0c9da1808c22