Groups, Unit Tests: a `creator_id` must be passed to `groups_create_group` to create a group.
Some of our unit tests were built to assume that they would always run with an authenticated user, but obviously this was incorrect. As a result, Groups were being created with `creator_id=0`, which was breaking other parts of the codebase, for certain tests.

This change also removes the `test_creating_new_group_as_anonymous_user` test that I added in r9118 which is now unnecessary.

See r9117, r9118, r9120 for other parts of this fix.
Fixes #5983 and the unit tests. :)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9121 cdf35c40-ae34-48e0-9cc9-0c9da1808c22