Delete users more selectively between tests.
In [7043] a workaround was introduced for a bug in the multisite tests. The
bug is that the admin user was not properly cleared after each test; the
workaround was to truncate the users table during setUp(). But this makes it
impossible to share user fixtures between tests. So in this changeset we
institute a narrower workaround that only deletes the specific problematic
user.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9138 cdf35c40-ae34-48e0-9cc9-0c9da1808c22