Properly highlight BuddyPress pages in a WordPress nav or page menu.
r8932 attempted to address how BuddyPress pages are highlighted in a
WordPress menu after the refactoring of `bp_core_load_template()` (see
r8821).  r8932 resulted in declaring the BuddyPress parent page as the
current page even when it wasn't.

This commit removes this declaration and changes the highlighting method
to rely on the WP nav and page menu filters.  This also fixes issues
with highlighting the BP parent page when the root profiles feature is
enabled (this previously did not work).

Fixes #5997.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9322 cdf35c40-ae34-48e0-9cc9-0c9da1808c22