Only create a single activity when a friendship is created and make sure to display it on each involved user personal activities.
When a friendship is created, a single public activity will be created. Thanks to the new `BP_Activity_Query` class introduced in version 2.2, we are able to fetch this activity on each involved user personal activity stream.
Once BuddyPress will be updated to 2.2, all `hide_sitewide` activities related to a created friendship will be removed as they are not necessary anymore.

Fixes #6040

props boonebgorges, r-a-y

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9326 cdf35c40-ae34-48e0-9cc9-0c9da1808c22