Activity: in `bp_has_activities()` use `bp_is_group()` instead of `groups_get_current_group()`
Since r9536, `bp_has_activities()` is using the  `groups_get_current_group()` function, which needs the Groups component to be active. There can be cases when this component has not been activated by the community administrator. In such cases, using this function is generating a fatal error. Using `bp_is_group()` is avoiding this error.

Fixes #6268

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9569 cdf35c40-ae34-48e0-9cc9-0c9da1808c22