Activity: improve blog post excerpts by showing rich media from the post.
Prior to this change, when we created excerpts for new/updated blog post activity items, we'd try to use any single image referenced in that blog post, and use it to decorate and bring interactivity to the activity stream item. This was inconsistently applied -- it was dependant on the length of the original post -- and ignored other possible media sources, such as oEmbeds and video/audio shortcodes.

Now, we use the media extractor (r9619) to extract comprehensive media information from a new/updated blog post, and use the information to build a much richer excerpt for blogs posts for activity items.

See #6177

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9621 cdf35c40-ae34-48e0-9cc9-0c9da1808c22