Introduce `bp_core_get_root_option()` and use throughout BP.
Previously, we referenced the `$bp->site_options` array directly. This causes
problems in cases where these options may be referenced before the array is
initially populated. `bp_core_get_root_option()` will fetch the requested
value from the array if it's been populated, and will populate it if it has not.

Fixes #6045.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9698 cdf35c40-ae34-48e0-9cc9-0c9da1808c22