XProfile: Introduce `BP_XProfile_Meta_Query` class.
This class extends `WP_Meta_Query` specifically for the Extended Profile component's shared metadata database table, which is unique in that it is used for metadata for field-groups, fields & field-options, and field-data for users.

The functionality of this class is not used in core yet, but will be used as part of #6347. Committing this in early to make iterating and testing easier during the 2.3 cycle.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9722 cdf35c40-ae34-48e0-9cc9-0c9da1808c22