Avatar UI: Add Backbone templates and adapt the legacy avatar templates.
The Backone templates will be in a new "assets" sub directory of `bp-templates\bp-legacy\buddypress`. For the introduction of the feature we are advising you *to not override* these specific templates from your theme. The underscore in each sub folder names of this "assets" directory is informing we wish to restrict the content of the sub folder to Core use only.

We are introducing a new template tag to load these Backbone templates : `bp_avatar_get_templates()`. We recommand that you update the avatar templates of your theme if you are overriding `groups/create`, `groups/single/admin` or `members/single/profile/change-avatar` once 2.3 will be released. If this new template tag is not into the template, then `bp_avatar_template_check()` will try to load them using template hooks.

Props johnjamesjacoby, hnla, r-a-y.

See #6290

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9757 cdf35c40-ae34-48e0-9cc9-0c9da1808c22